requirements.txt
numpy
pandas
tweepy
matplotlib
seaborn
textblob
mysql-connector-python
geopy
plotly
nltk
dash
plotly
dash
plotly
pandas
flask
psycopg2